+++ src/main/java/app/habitzl/elasticsearch/status/monitor/tool/configuration/DefaultConfigurationLoader.java
@@ -143,7 +143,7 @@ public class DefaultConfigurationLoader implements ConfigurationLoader {
 
     private void printHelp() {
         String cmdLineSyntax = "java -jar elasticsearch-status-monitor.jar";
-        String header = "This tool provides a quick overview of an Elasticsearch cluster and analyses the data for potential problems.";
+        String header = "This tool provides an overview of an Elasticsearch cluster and analyses the data for potential problems.";
         String footer = "Please report issues at: https://github.com/romanha/elasticsearch-status-monitor/issues";
 
         HelpFormatter helpFormatter = new HelpFormatter();